backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers
djoser
drf-yasg==1.21.7
django_filter
gunicorn==20.1.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
python-decouple
pytz==2024.1
PyYAML==6.0.1
qrcode
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
psycopg2-binary